requirements.txt
cx_Oracle
google-generativeai
pandas
openpyxl
nepali-datetime
flask
joblib
python-dotenv
python